requirements.txt
yfinance
streamlit==1.8.1
pandas==1.4.2
Pillow==9.1.0
dynaconf==3.1.7
urllib
